update requirement 8, a new class added, general corrections, comments added

+++ NeoTunes/src/model/Song.java
@@ -43,7 +43,7 @@ public class Song extends Audio {
         return sales;
     }
 
-    public void setSales(int sales) {
+    public void setSales(double sales) {
         this.sales = sales;
     }
